Transaction.from_sentry_trace: Correctly set sampled, fix edge case crash
Previously this method didn't set the @sampled variable on the created
transaction, causing subsequent child spans to not be sent, even if
the parent span was sampled.

This was caused by a missing call to set_initial_sample_desicion in this
code path.

In passing, fix a crash in this method when an invalid value is passed
that doesn't match the header regexp.